Renamed template .gitignore file to gitignore so that npm will not exclude it during packaging (see npm/npm#1862).
Per the discussion on the npm issue, if there is no .npmignore file but there is a .gitignore file, it will be included in the npm package as .npmignore. If there is an .npmignore file and also a .gitignore file, the .gitignore file will simply be excluded from the package.